mems-ana_demo/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mems-ana"
version = "0.1.0"
description = "Lightweight MEMS analysis (pre-FEM): schematic piezo/ferroelectric loops and visualization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "matplotlib>=3.7",
  "pillow>=10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
